+++ config/basic_auth.php
@@ -16,4 +16,17 @@ return [
     'password' => env('BASIC_AUTH_PASSWORD', ''),
     'username' => env('BASIC_AUTH_USERNAME', ''),
 
+    /*
+    |--------------------------------------------------------------------------
+    | Basic Authentication Settings
+    |--------------------------------------------------------------------------
+    |
+    | This options configure the middleware behaviour.
+    |
+    */
+
+    'alias' => env('BASIC_AUTH_ALIAS', 'basic.auth'),
+    'global' => env('BASIC_AUTH_GLOBAL', false),
+    'group' => env('BASIC_AUTH_GROUP', 'web'),
+
 ];